Authentication

+++ src/features/tickets/actions/delete-ticket.ts
@@ -1,15 +1,30 @@
-"use server";
+'use server';
 
-import { redirect } from "next/navigation";
-import { setCookieByKey } from "@/actions/cookies";
-import { fromErrorToActionState } from "@/components/form/utils/to-action-state";
-import { prisma } from "@/lib/prisma";
-import { ticketsPath } from "@/paths";
+import { redirect } from 'next/navigation';
+import { setCookieByKey } from '@/actions/cookies';
+import {
+  fromErrorToActionState,
+  toActionState,
+} from '@/components/form/utils/to-action-state';
+import { getAuthOrRedirect } from '@/features/auth/queries/get-auth-or-redirect';
+import { isOwner } from '@/features/auth/utils/is-owner';
+import { prisma } from '@/lib/prisma';
+import { ticketsPath } from '@/paths';
 
 export const deleteTicket = async (id: string) => {
-  await new Promise((resolve) => setTimeout(resolve, 2000));
+  const { user } = await getAuthOrRedirect();
 
   try {
+    const ticket = await prisma.ticket.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (!ticket || !isOwner(user, ticket)) {
+      return toActionState('ERROR', 'Not authorized');
+    }
+
     await prisma.ticket.delete({
       where: {
         id,
@@ -19,6 +34,6 @@ export const deleteTicket = async (id: string) => {
     return fromErrorToActionState(error);
   }
 
-  await setCookieByKey("toast", "Ticket deleted");
+  await setCookieByKey('toast', 'Ticket deleted');
   redirect(ticketsPath());
 };
